written out the meal routine. GO TEST

+++ include/philo.h
@@ -6,7 +6,7 @@
 /*   By: mintan <mintan@student.42singapore.sg>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/07 10:49:46 by mintan            #+#    #+#             */
-/*   Updated: 2024/12/19 09:54:18 by mintan           ###   ########.fr       */
+/*   Updated: 2024/12/19 17:17:49 by mintan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,8 +82,7 @@ typedef struct s_philo
 	int				full;
 	int				r_no;
 	int				l_no;
-	// int				state;
-	int				eat_times;
+	long			eat_times;
 	long			ms_last_eat;
 	pthread_mutex_t	*r_fork;
 	pthread_mutex_t	*l_fork;
